Check if folder exists before deleting and increase the timeout for Logs to be created

+++ Deploy/service/impl/ProjectServiceImpl.cs
@@ -35,7 +35,8 @@ public class ProjectServiceImpl : ProjectService
         if (string.IsNullOrEmpty(_configRepository.GetSystemConfig().ProjectCacheFolderToDelete)) return;
 
         var folderToDelete = Path.Combine(projectPath, _configRepository.GetSystemConfig().ProjectCacheFolderToDelete!);
-        FileSystem.DeleteDirectory(folderToDelete, UIOption.OnlyErrorDialogs, RecycleOption.SendToRecycleBin);
+        if (FileSystem.DirectoryExists(folderToDelete))
+            FileSystem.DeleteDirectory(folderToDelete, UIOption.OnlyErrorDialogs, RecycleOption.SendToRecycleBin);
     }
 
     private static void BuildProject(string projectPath) => ExecUtils.RunCommand(projectPath, GradleBuildCommand);
